split client channel
