+++ agent/package.json
@@ -19,6 +19,7 @@
         "@ai16z/client-discord": "workspace:*",
         "@ai16z/client-telegram": "workspace:*",
         "@ai16z/client-twitter": "workspace:*",
+        "@ai16z/client-terminal": "workspace:*",
         "@ai16z/eliza": "workspace:*",
         "@ai16z/plugin-bootstrap": "workspace:*",
         "@ai16z/plugin-conflux": "workspace:*",
