+++ agent/src/libs/utils.ts
@@ -1,0 +1,68 @@
+import {
+    settings,
+    Character,
+    ModelProviderName,
+    // elizaLogger,
+} from "@ai16z/eliza";
+
+export const wait = (minTime: number = 1000, maxTime: number = 3000) => {
+    const waitTime =
+        Math.floor(Math.random() * (maxTime - minTime + 1)) + minTime;
+    return new Promise((resolve) => setTimeout(resolve, waitTime));
+};
+
+export function getTokenForProvider(
+    provider: ModelProviderName,
+    character: Character
+) {
+    switch (provider) {
+        case ModelProviderName.OPENAI:
+            return (
+                character.settings?.secrets?.OPENAI_API_KEY ||
+                settings.OPENAI_API_KEY
+            );
+        case ModelProviderName.LLAMACLOUD:
+            return (
+                character.settings?.secrets?.LLAMACLOUD_API_KEY ||
+                settings.LLAMACLOUD_API_KEY ||
+                character.settings?.secrets?.TOGETHER_API_KEY ||
+                settings.TOGETHER_API_KEY ||
+                character.settings?.secrets?.XAI_API_KEY ||
+                settings.XAI_API_KEY ||
+                character.settings?.secrets?.OPENAI_API_KEY ||
+                settings.OPENAI_API_KEY
+            );
+        case ModelProviderName.ANTHROPIC:
+            return (
+                character.settings?.secrets?.ANTHROPIC_API_KEY ||
+                character.settings?.secrets?.CLAUDE_API_KEY ||
+                settings.ANTHROPIC_API_KEY ||
+                settings.CLAUDE_API_KEY
+            );
+        case ModelProviderName.REDPILL:
+            return (
+                character.settings?.secrets?.REDPILL_API_KEY ||
+                settings.REDPILL_API_KEY
+            );
+        case ModelProviderName.OPENROUTER:
+            return (
+                character.settings?.secrets?.OPENROUTER ||
+                settings.OPENROUTER_API_KEY
+            );
+        case ModelProviderName.GROK:
+            return (
+                character.settings?.secrets?.GROK_API_KEY ||
+                settings.GROK_API_KEY
+            );
+        case ModelProviderName.HEURIST:
+            return (
+                character.settings?.secrets?.HEURIST_API_KEY ||
+                settings.HEURIST_API_KEY
+            );
+        case ModelProviderName.GROQ:
+            return (
+                character.settings?.secrets?.GROQ_API_KEY ||
+                settings.GROQ_API_KEY
+            );
+    }
+}
