+++ characters/tate.character.json
@@ -1,7 +1,7 @@
 {
     "name": "tate",
-    "clients": [],
-    "modelProvider": "anthropic",
+    "clients": ["direct", "terminal"],
+    "modelProvider": "openai",
     "settings": {
         "secrets": {},
         "voice": {
