+++ characters/trump.character.json
@@ -1,6 +1,6 @@
 {
     "name": "trump",
-    "clients": [],
+    "clients": ["direct", "terminal"],
     "modelProvider": "openai",
     "settings": {
         "secrets": {},
