+++ packages/core/src/types.ts
@@ -590,6 +590,7 @@ export enum Clients {
     DIRECT = "direct",
     TWITTER = "twitter",
     TELEGRAM = "telegram",
+    TERMINAL = "terminal",
 }
 /**
  * Configuration for an agent character
